Added extra axis space
The main dimension of the axes are calculated purly based on the rendered text of a tick.  If the user wishes to transform or add a dy to the axis, it can cause clipping of the SVG text elements.  This change allows users to add extra space to the primary dimension of each axis.